pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "softkill9000"
version = "1.0.0"
description = "Multi-Agent Cosmic Mission Simulator with Ethics-Aware Reinforcement Learning"
authors = [
    {name = "MotionBlendAI Team", email = "team@motionblendai.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Visualization",
]
keywords = [
    "reinforcement-learning",
    "multi-agent",
    "simulation",
    "motion-capture",
    "visualization",
    "ethics-aware-ai"
]

dependencies = [
    "numpy>=1.24.0,<2.0.0",
    "matplotlib>=3.7.0,<4.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pyyaml>=6.0.0,<7.0.0",
]

[project.optional-dependencies]
gradio = [
    "gradio>=4.0.0,<5.0.0",
    "imageio>=2.31.0,<3.0.0",
    "gTTS>=2.3.0,<3.0.0",
]
api = [
    "fastapi>=0.104.0,<1.0.0",
    "uvicorn[standard]>=0.24.0,<1.0.0",
]
dev = [
    "pytest>=7.4.0,<8.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "black>=23.7.0,<24.0.0",
    "flake8>=6.1.0,<7.0.0",
    "mypy>=1.5.0,<2.0.0",
    "isort>=5.12.0,<6.0.0",
]
docs = [
    "mkdocs>=1.5.0,<2.0.0",
    "mkdocs-material>=9.4.0,<10.0.0",
    "mkdocstrings[python]>=0.23.0,<1.0.0",
]
notebooks = [
    "jupyter>=1.0.0,<2.0.0",
    "ipykernel>=6.25.0,<7.0.0",
]
all = [
    "softkill9000[gradio,api,dev,docs,notebooks]",
]

[project.urls]
Homepage = "https://github.com/BkAsDrP/Softkill9000"
Documentation = "https://github.com/BkAsDrP/Softkill9000/docs"
Repository = "https://github.com/BkAsDrP/Softkill9000"
Issues = "https://github.com/BkAsDrP/Softkill9000/issues"

[project.scripts]
softkill9000 = "softkill9000.cli:main"
softkill9000-api = "softkill9000.api.server:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["softkill9000*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=softkill9000",
    "--cov-report=html",
    "--cov-report=term-missing",
]
